+++ gui/include/graphics/scene_camera_2d.hpp
@@ -13,7 +13,7 @@ class SceneCamera2d : public SceneCamera {
 
     bool handleMouseButton(int button, bool down) override;
     bool handleScroll(double offset) override;
-    bool handleMouseMoved(float x, float y) override;
+    bool handleMouseMoved(double x, double y) override;
     bool handleKey(int key, bool down, int mods) override;
 
     glm::vec3& position() override { return position_; }
@@ -26,8 +26,8 @@ class SceneCamera2d : public SceneCamera {
     float angle_ = 0.0f;
     float scale_ = 0.5f;
 
-    float prev_x_ = -1.1f;
-    float prev_y_ = -1.1f;
+    double prev_x_ = -1.1;
+    double prev_y_ = -1.1;
 
     bool dragging_ = false;
 };
