+++ gui/include/input_handler.hpp
@@ -8,7 +8,8 @@ class InputHandler {
     virtual bool handleScroll(double /* offset */) { return false; }
     virtual bool handleKey(int /* key */, bool /* down */, int /* mods */) { return false; }
     virtual bool handleChar(unsigned int /* c */) { return false; }
-    virtual bool handleMouseMoved(float /* x */, float /* y */) { return false; }
+    // glfwGetCursorPos takes double as arguments
+    virtual bool handleMouseMoved(double /* x */, double /* y */) { return false; }
 
     virtual int priority() const { return 10; }
 };
