+++ gui/include/scene.hpp
@@ -67,7 +67,7 @@ class SceneList : public RenderTarget,
 
     bool handleMouseButton(int button, bool down) override;
     bool handleScroll(double offset) override;
-    bool handleMouseMoved(float x, float y) override;
+    bool handleMouseMoved(double x, double y) override;
     bool handleKey(int key, bool down, int mods) override;
 
     void handle(tomop::Packet& packet) override { send(packet); }
